Add Qwen3-VL support with deepstack visual embeds
- Register qwen3_vl_vision encoder and qwen3_vl projector/get_embeds
  in utils/const.py and class resolvers in utils/utils.py.
- Add new encoder (configuration/modeling_qwen3vl_vision.py) and
  projector (projector_qwen3_vl_patchmerger.py) modules.
- Extend PatchMergerProjectorConfig with qwen3_vl model_type and
  deepstack_visual_indexes field.
- Skip window-attention path in Qwen2_5VLPreEncoder and
  Qwen2_5VLGetEmbeds when window_size == 0 (Qwen3-VL case).
- Qwen2VLPreLLM: emit visual_pos_masks (with overture padding) for
  deepstack injection.
- FloatPipeline: precompute Qwen3-VL 2D positional embeddings,
  propagate vision rotary embeddings across encoder chunks, capture
  deepstack intermediates, run postshuffle mergers, and add deepstack
  embeds at image-token positions in the first N decoder layers (with
  mask alignment to padded prefill).
- QuantizedPipeline / inference_quantized.py: filter encoder model
  files by .tflite/.mlir extension instead of excluding .json.
- Qwen2VLGetEmbeds: cast image embeds to text-embed dtype before
  scattering to avoid dtype mismatch.

+++ models/get_embeds/qwen2_vl.py
@@ -72,7 +72,7 @@ class Qwen2VLGetEmbeds(BaseGetEmbedsHook):
             image_embeds = torch.from_numpy(image_embeds).to(
                 device=self.text_embedding_layer.weight.device, dtype=text_embeds.dtype
             )
-        text_embeds[selected] = image_embeds.reshape(-1, c)
+        text_embeds[selected] = image_embeds.reshape(-1, c).to(dtype=text_embeds.dtype)
         text_embeds = text_embeds.reshape(b, n, c)
         kwargs.update({'text_tokens': text_tokens})
